:recycle: Merge duplicate travel desk edit tables.
